Renaming of indexes
