+++ src/ts/components/formats/rgb.ts
@@ -1,5 +1,5 @@
-import { Phrases } from "../types/SectionPhrases";
-import randomNumberFromInterval from "./randomNumberFromInterval";
+import { Phrases } from "../../types/SectionPhrases";
+import randomNumberFromInterval from "../randomNumberFromInterval";
 
 export default class RGB {
   name: string;
